requirements.txt
Django==4.0.3
djangorestframework==3.13.1
gunicorn==20.1.0
psycopg2==2.9.3